Add folder __tests__/__utils__
So we have the same structure for test helper as in
https://github.com/serlo/api.serlo.org/
